move register to login-form folder
rename and edit imports to reflect changes
